+++ src/xenia/kernel/util/presence_string_builder.cc
@@ -10,6 +10,8 @@
 #include "xenia/kernel/util/presence_string_builder.h"
 #include "third_party/fmt/include/fmt/format.h"
 
+#include "xenia/base/string.h"
+
 namespace xe {
 namespace kernel {
 namespace util {
