+++ src/xenia/kernel/util/xlast.h
@@ -11,6 +11,7 @@
 #define XENIA_KERNEL_UTIL_XLAST_H_
 
 #include <map>
+#include <memory>
 #include <optional>
 #include <string>
 #include <vector>
